Finally included games as package data
This is probably not the right way, but the games/ subdirectory is now a
Python package so include_package_data will include its contents when using
recursive_include in MANIFEST.in. The code simply imports games and looks at
games.__file__ to determine where the game data is located. It works, but
Python packaging is still an unclear mess!